requirements.txt
streamlit
langchain
langchain-google-genai
langchain-community
faiss-cpu
python-dotenv
pypdf